requirements.txt
# booktran v0.1 - 依赖清单

# 文档解析
PyMuPDF>=1.23.0          # PDF 解析
ebooklib>=0.18           # EPUB 解析
python-docx>=1.0.0       # DOCX 解析
beautifulsoup4>=4.12.0   # HTML 解析
lxml>=4.9.0              # HTML/XML 解析
chardet>=5.0.0           # 编码检测

# 图片处理
Pillow>=10.0.0           # 图片处理

# HTTP 请求
httpx>=0.25.0            # 异步 HTTP 客户端

# 配置
PyYAML>=6.0              # YAML 配置解析

# EPUB 构建
ebooklib>=0.18           # EPUB 生成
